Removed /tictactoe leaderboard
